requirements.txt
python-telegram-bot
telethon
logging
asyncio
